Refactor : toEntity 컨버터로 분리
